added a tester ... to test lib_ft

+++ Libft/ft_substr.c
@@ -8,7 +8,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 
 	if (!s)
 		return (NULL);
-	if (start >= ft_strlen(s))
+	if (start >= (unsigned int)ft_strlen(s))
 		return (ft_strdup(""));
 	str = (char *)malloc(sizeof(char) * (len + 1));
 	if (str == NULL)
